Add 32 bit address handling for mappers

+++ src/mapper.rs
@@ -2,7 +2,7 @@ pub mod nrom;
 pub mod uxrom;
 
 pub trait Mapper {
-    fn map_prg(&self, address: u16) -> u16;
-    fn map_chr(&self, address: u16) -> u16;
+    fn map_prg(&self, address: u16) -> u32;
+    fn map_chr(&self, address: u16) -> u32;
     fn bank_select(&mut self, value: u8);
 }
